refactor: split the form logic to useBookForm hook.
