+++ src/components/BookingsLayer.tsx
@@ -100,7 +100,12 @@ const BookedBlock = ({
         // only staff or booked student can review/edit a booking.
         // TEMP: Permission check disabled until backend returns `intra` and `role`
         // if (user?.role !== 'staff' && !isCurrUser) return;
-        setFormProp({ booking: slot, roomId: roomId, startTime: new Date(slot.start) });
+        setFormProp({
+          booking: slot,
+          roomId: roomId,
+          startTime: new Date(slot.start),
+          channel: 'sheet',
+        });
       }}
     >
       <span
